rescan-scsi-bus.sh: Use "seq -s"
Since 8c86fe2 ("library: add to '.so' name; rescan-scsi-bus: multiple patches
to sync with Suse", we correctly use quoting when we print $idsearch or
$lunsearch. This causes a change in the output of the script, because
these number lists are generated by the GNU seq utility, which uses "\n" as
default separator, and the quoting now causes the newlines to be preserved.
Use "seq -s" to generate space-separated lists instead.