rescan-scsi-bus: Add option --no_lip_scan to scan FC Host without lip
lip is a FC operation and can be used to rescan device. While it can not
specify the channel and target. So here add a new option make scan FC
hosts according SCSI general interface which is
/sys/class/scsi_host/hostX/scan.

The default value of no_lip_scan is 0 which do not change the behavior
default.

Fix https://github.com/doug-gilbert/sg3_utils/issues/20

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>