rescan-scsi-bus.sh: add option --no-lip-scan
lip is a FC operation and can be used to rescan device but it can not
specify the channel and target. So add a new option to scan FC hosts
according to the SCSI general interface in sysfs:
/sys/class/scsi_host/hostX/scan.

The default value of no_lip_scan is 0 which does not change the
default behavior.

Fix https://github.com/doug-gilbert/sg3_utils/issues/20

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>


git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@967 6180dd3e-e324-4e3e-922d-17de1ae2f315