sg3_utils: sg_sat_datetime command to get/set the date & time on ATA devices.
Similar to sg_timestamp for SCSI, sg_sat_datetime allows the
date and time (timestamp) to be set-on and returned-from
ATA devices.  This provides the ability to syncronize the
system clock with the device.  The SET DATE & TIME EXT and
READ LOG [DMA] EXT ATA commands are sent using ATA
PASSTHROUGH SCSI commands to set and retrieve the date and
time on the device.