udev rules: allow command line overrides for ID_SERIAL and symlink creation
commit d7b8da0 ("udev rules: restrict use of ambiguous device IDs patchset")
changes device identification and symlink creation for SCSI devices.

In the worst case, these changes can cause devices to fail to be detected, and
may even cause boot failure. While this is unlikely, it may be helpful to add
kernel command line parameters to override the settings as a means for
emergency recovery.

The parameters are "udev.scsi_id_serial_src=..." and
"udev.scsi_symlink_src=... ", where the value is any combination of the
letters "LTVS" for NAA *L*ocal, *T*10 vendor ID, *V*endor-specific, and
*S*erial-number based identification, respectively.

Signed-off-by: Martin Wilck <mwilck@suse.com>